:white_check_mark: Verifiy Email Component Done
