--- react/acm-chapter-website/src/pages/verifiyEmail/verifyEmail.scss
@@ -1,4 +1,0 @@
-.verifyEmail {
-  top: 100px;
-  left: 100px;
-}
