+++ react/acm-chapter-website/src/pages/verifyEmail/verifyEmail.scss
@@ -1,0 +1,44 @@
+.login {
+  background: var(--background);
+  text-align: center;
+  display: flex;
+  align-items: center;
+
+  &__landing{
+      background: var(--background);
+      box-shadow: rgba(100, 100, 111, 0.2) 0px 7px 29px 0px;
+      overflow-y: hidden;
+      button{
+        font-size: 20px;
+        min-width: 150px;
+
+      }
+
+      header {
+        margin-top: 0;
+        width: 75%;
+
+        h2 {
+          margin-top: 0;
+          font-size: 2.5rem;
+        }
+        p {
+          margin-top: 0;
+          color: var(--secondary);
+        }
+      }
+
+      p {
+        margin-top: 3%;
+        font-size: 2.5rem;
+        color: brown;
+      }
+
+      .verified {
+        color: rgb(48, 172, 32);
+        margin-bottom: 5vh;
+      }
+
+    }
+
+  }
